# docs/v1/swagger.yaml
basePath: /api/v1
definitions:
  dto.UserCreateRequest:
    properties:
      email:
        description: |-
          Email of the user
          Required: true
          Example: john.doe@example.com
        type: string
      first_name:
        description: |-
          FirstName of the user
          Required: true
          Example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        description: |-
          LastName of the user
          Required: true
          Example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        description: |-
          Password for the user account
          Required: true
          Example: StrongP@ssw0rd
        minLength: 8
        type: string
      second_name:
        description: |-
          SecondName of the user
          Example: Michael
        maxLength: 100
        minLength: 2
        type: string
      user_type:
        description: |-
          UserType defines the role of the user
          Enum: user, admin
          Example: user
        enum:
        - user
        - admin
        type: string
      username:
        description: |-
          Username of the user
          Required: true
          Example: johndoe123
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  dto.UserUpdateRequest:
    properties:
      email:
        description: |-
          Email of the user
          Example: john.doe@example.com
        type: string
      first_name:
        description: |-
          FirstName of the user
          Example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        description: |-
          LastName of the user
          Example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        description: |-
          Password for the user account
          Example: StrongP@ssw0rd
        minLength: 8
        type: string
      second_name:
        description: |-
          SecondName of the user
          Example: Michael
        maxLength: 100
        minLength: 2
        type: string
      user_type:
        description: |-
          UserType defines the role of the user
          Enum: user, admin
          Example: user
        enum:
        - user
        - admin
        type: string
      username:
        description: |-
          Username of the user
          Example: johndoe123
        maxLength: 50
        minLength: 3
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
      request_id:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Go Platform Template - Base platform with Auth, User, File, and RBAC
    features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Platform Template API
  version: "1.0"
paths:
...
  /users:
    get:
      parameters:
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Limit for pagination
        in: query
        name: limit
        type: integer
      - description: Filter by username
        in: query
        name: username
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by user type
        in: query
        name: user_type
        type: string
      - description: Sort by field (created_at or username)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users with pagination, filters, and sorting
      tags:
      - Users
  /users/:
    post:
      consumes:
      - application/json
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"